add: 결제 승인 테스트 코드 작성

+++ src/main/kotlin/com/yh/paymentwebflux/payment/domain/Product.kt
@@ -4,7 +4,7 @@ import java.math.BigDecimal
 
 data class Product (
     val id: Long,
-    val amount: BigDecimal,
+    val amount: Long,
     val quantity: Int,
     val name: String,
     val sellerId: Long
